# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HEXA Purchase Assistant API
  version: 0.1.0
  description: |
    FastAPI backend for Flutter app, WhatsApp webhook, and Super Admin.
    AI is used for parsing only; business logic stays server-side.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Create account (username, email, password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: JWT access + refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '409':
          description: Email or username already registered
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Sign in with email or username + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT access + refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid credentials
  /v1/auth/google:
    post:
      tags: [auth]
      summary: Sign in with Google (verify ID token, create or link user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: JWT access + refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid Google ID token
        '503':
          description: Google auth not configured on server
  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 64
          description: Lowercase letters, digits, underscore
        password:
          type: string
          minLength: 8
          maxLength: 128
    LoginRequest:
      type: object
      required: [email_or_username, password]
      properties:
        email_or_username:
          type: string
          description: Full email or username (no @ required for username)
        password:
          type: string
    GoogleAuthRequest:
      type: object
      required: [id_token]
      properties:
        id_token:
          type: string
          description: Google Sign-In ID token (JWT)
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
...